requirements.txt
altgraph==0.17
APScheduler==3.6.3
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
click==7.1.1
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
distro==1.5.0
entrypoints==0.3
Flask==1.1.2
future==0.18.2
importlib-metadata==1.6.0
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.2.0
nbconvert==5.6.1
nbformat==5.0.5
notebook==6.0.3
packaging==20.3
pandocfilters==1.4.2
parso==0.7.0
pefile==2019.4.18
pickleshare==0.7.5
pip-custom-platform==0.5.0
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
py==1.8.1
Pygments==2.6.1
PyInstaller==3.6
pylint==2.4.4
pymonkey==0.3.1
pyparsing==2.4.7
line-bot-sdk==1.16.0

